separate out the SensorManager interfaces
