+++ src/app.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 import { App } from "./components/app";
-import { SensorConnectorManager } from "./models/sensor-manager";
+import { SensorConnectorManager } from "./models/sensor-connector-manager";
 
 let sensorManager = new SensorConnectorManager();
